refactor: material registry splitted in hpp and cpp
